+++ tests/test_time.py
@@ -28,3 +28,29 @@ class TestTime(unittest.TestCase):
             self.assertGreaterEqual(measurement(), 0.5)
         measured()
 
+    def test_measure_method_decorator(self):
+        class Test:
+            @measure()
+            def try_measure(self, measurement):
+                time.sleep(0.5)
+                return measurement()
+
+        self.assertGreaterEqual(Test().try_measure(), 0.5)
+
+    def test_resuming(self):
+        a = measure()
+        time.sleep(0.5)
+        self.assertGreaterEqual(a(), 0.5)
+        a.stop()
+        time.sleep(0.5)
+        self.assertLess(a(), 0.7)
+        a.start()
+        time.sleep(0.5)
+        self.assertLess(a(), 1.2)
+        self.assertRaises(TypeError, lambda: a.reset())
+        a.stop()
+        self.assertRaises(TypeError, lambda: a.stop())
+        a.reset()
+        time.sleep(0.5)
+        self.assertGreaterEqual(a(), 0.5)
+        self.assertLess(a(), 0.7)