+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * added `CallNoOftenThan`
 * added `install_force_gc_collect`
+* added comparison and arithmetics options to `AtomicNumber`