+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * added `CallNoOftenThan`
 * added `install_force_gc_collect`
 * added comparison and arithmetics options to `AtomicNumber`
+* added `LabeledMetric`