+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 * added `install_force_gc_collect`
 * added comparison and arithmetics options to `AtomicNumber`
 * added `LabeledMetric`
+* refactored `MeasurableMixin`