+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # v2.7.27
 
+* made `WouldWaitMore` be subclass of `TimeoutError`
+* made `ConfigurationError` be subclass of `ValueError`