+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # v2.7.29
 
 * added `wait_until_equal` to `AtomicNumber`
+* `measure` will behave correctly when passed a method
+* fixed rules regarding re-using `measure`
+* added `auto_adapt_to_methods`