+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 * added `wait_until_equal` to `AtomicNumber`
 * `measure` will behave correctly when passed a method
 * fixed rules regarding re-using `measure`
+* added new option to `measure` - `create_stopped`
 * added `auto_adapt_to_methods`