+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-# v2.7.30
-
-* fixed semantics of `measure` with decorators
+# v2.7.31